added node class
Node class can be defined for selection of a (remote) node to run the operation. It will check the nodes of the node class and the first node which is active and has the file system mounted will be selected. If the local node is member of the node class it will be checked first.